python/requirements.txt
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
yfinance==0.2.37
requests==2.31.0
beautifulsoup4==4.12.3
textblob==0.18.0.post0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
